train/requirements.txt
tensorflow==2.5.0
pandas==1.3.5
scipy==1.7.3
tsaug==0.2.1
sklearn==0.0
matplotlib==3.5.1
seaborn==0.11.2
pyserial==3.5